debug

+++ mymodels/dataset.py
@@ -242,12 +242,12 @@ def collate_fn(batch):
     mask = np.stack(mask, axis=0)
     seq = [s for s in seq]
 
-    res2d = torch.from_numpy(res2d).double()
-    i3d = torch.from_numpy(i3d).double()
-    relation = torch.from_numpy(relation).double()
-    object_ = torch.from_numpy(object_).double()
-    res2d_mask = torch.from_numpy(res2d_mask).double()
-    i3d_mask = torch.from_numpy(i3d_mask).double()
+    res2d = torch.from_numpy(res2d).float()
+    i3d = torch.from_numpy(i3d).float()
+    relation = torch.from_numpy(relation).float()
+    object_ = torch.from_numpy(object_).float()
+    res2d_mask = torch.from_numpy(res2d_mask).float()
+    i3d_mask = torch.from_numpy(i3d_mask).float()
     numberic = torch.from_numpy(numberic)
     mask = torch.from_numpy(mask)
